+++ crates/wdl-engine/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 * Implemented remote file localization for task execution (#[386](https://github.com/stjude-rust-labs/wdl/pull/386)).
+* Implemented concurrent file downloads for localization for task execution (#[424](https://github.com/stjude-rust-labs/wdl/pull/424)).
 
 #### Fixed